Opponents Shoot Player
Opponents will now shoot a player if they are in their direct line of
fire. I also changed the start location of tank shells so that instead
of using an arbitrary number the distance from the firing tank
immediately after a shell is fired is now proportional to the size of the firing tank.