Added Map Scalling
The map now scales to fill the whole screen (or a specified portion of the screen).
The map will remain a constant size, no matter the complexity of the maze (e.g. 2x2 or 16x9, etc.)

I intend to add code that will keep the width of all walls constant, as currently the scale differently based on the x and y scale.

I also fixed an issue where the map objects were being placed overlapping one another.