Refactored MapExits
MapExit has been refactored.

The class MapExit has been removed, and its behaviour has now been placed into the MapObject class.

As a result all code refering to MapExits has been greatly simplified as instead of requiring different functions,  a mapObject is instead tested to see if it is a mapExit, and if it is, it will exhibit the same behaviour as the previous MapExit class.