Created Button and UIScreen classes
Created a Button class which when pressed will perform an operation (Currently just printing a message to the terminal). The button has a number of states depending on if it is being pressed or if the mouse is over the button.

I also created a UIScreen abstract class which will be extended when creating new UI Screens. The class contains a number of operations for adding buttons, updating the buttons, etc.

To this end, I created a MainMenu class which currently houses some buttons (using test textures), in the layout of the main menu.

Note: Buttons are not complete in their functionality.

Different buttons will either be different classes, or when creating a button a function from the screen they are being displayed on will be passed in, and that function will determine their functionality