Added a Proper Pause Screen
Added a proper pause screen (i.e. UI Screen), that has buttons for quitting the game and continuing.

It appears when the user presses escape, and works on both the UIScreen and gameplay scenes.

The classes PauseScreen and ResumeButton were created in order for the pause screen to function correctly