Fixed Pervasive UI Bug
Fixed a Bug where the UI would not respond properly if the window was re-scalled past a certain point.

This fix now means that the UI will detect the mouse properly no matter the window size.

I have also disabled the custom mouse cursor in this commit, as it has some issues with window scalling and being slightly off the actual mouse's position